# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Filmes",
        "description": "Documentação da API para gerenciamento de filmes",
        "version": "1.0.0"
    },
    "paths": {
        "/api/movies": {
            "get": {
                "tags": [
                    "Filmes"
                ],
                "summary": "Lista todos os filmes",
                "operationId": "a7859dc5cd106bbeae5492e22df893ec",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Título do filme",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Ano de lançamento",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "director",
                        "in": "query",
                        "description": "Nome do diretor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Movie"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Filmes"
                ],
                "summary": "Exibe os detalhes de um filme",
                "operationId": "a7031f1ec2eb856acbdd7f7054ed001d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do filme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "title": "Filme",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "O Poderoso Chefão"
                    },
                    "year": {
                        "type": "integer",
                        "example": 1972
                    },
                    "director": {
                        "type": "string",
                        "example": "Francis Ford Coppola"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Filmes",
            "description": "Filmes"
        }
    ]
}